Interactive data cleaning

Interactive data cleaning

dcr_app(dframe, browser = TRUE)

Arguments

dframe
dataframe, tbl or data.table with data for viewing and/or cleaning

browser
logical, should app start in OS's default browser? (default TRUE)

Details

datacleanr provides an interactive data overview, and allows reproducible subsetting and (manual, interactive) visual outlier detection and annotation. The app works with data.frame, tibble and data.table objects. Note, that spatial data is supported; this requires that lon and lat columns in degrees are present. The interactive visualization relies on subsetting (i.e. highlighting) sub-groups of the provided data. These groups are composed of unique combinations of columns in the data set (must be factor), and are carried through the app for exploratory analyses (tab Overview), and plotting (tab Visualization). These groups should ideally be chosen to facilitate a convenient viewing and cleaning process. For example, a data set with time series of multiple sensors could be grouped by sensor and/or additional columns, such that periods of interest can be visualized and cleaned simultaneously in the interactive plot. The tab Extraction provides code to reproduce the entire procedure, which can be copied, or sent directly to an active RStudio script.