Enforce simple column types

Analyse each column of a data.frame and try to convert character columns to more appropriate types (numeric, integer or factor).

Usage

enforce_types(data, num_threshold = 0.9, max_factor_levels = 20)

Arguments

data

The input data.frame.

num_threshold

Proportion (0 to 1). Minimum proportion of valid numeric values required to convert to numeric (default 0.9).

max_factor_levels

Integer. Maximum number of unique values to allow conversion to a factor (default 20).

Value

A new data.frame with optimised column types.

Details

The algorithm processes the data column by column:

Ignores columns that are already numeric, factors or dates.

Trims surrounding whitespace with trimws.

Numeric conversion: If the proportion of values that can be converted to numeric exceeds num_threshold, the column is converted to numeric (or integer when possible).

Factor conversion: Otherwise, if the number of unique values is less than or equal to max_factor_levels, the column is converted to a factor.

Otherwise, the column remains as character.

See also

Examples

df <- data.frame(
  id = c("1", "2", "3"),              # Should become integer
  cat = c("A", "A", "B"),             # Should become factor
  text = c("Unique1", "Unique2", "Unique3"), # Remains character
  stringsAsFactors = FALSE
)

str(df) # All columns are character initially
#> 'data.frame':	3 obs. of  3 variables:
#>  $ id  : chr  "1" "2" "3"
#>  $ cat : chr  "A" "A" "B"
#>  $ text: chr  "Unique1" "Unique2" "Unique3"

df_typed <- enforce_types(df, max_factor_levels = 2)
str(df_typed) # Types corrected
#> 'data.frame':	3 obs. of  3 variables:
#>  $ id  : int  1 2 3
#>  $ cat : Factor w/ 2 levels "A","B": 1 1 2
#>  $ text: chr  "Unique1" "Unique2" "Unique3"