Server Module: Selection Annotator

Server Module: Selection Annotator

module_server_extract_code(
  input,
  output,
  session,
  df_label,
  filter_df,
  gvar,
  statements,
  sel_points,
  overwrite,
  is_on_disk,
  out_path
)

Arguments

input, output, session
standard shiny boilerplate

df_label
string, name of original df input

filter_df
reactiveValue data frame with filter statements and scoping lvl

gvar
reactive character, grouping vars for dplyr::group_by

statements
reactive, lgl, vector of working statements

sel_points
reactiveValue, data frame with selected point keys, annotations, and selection count

overwrite
reacive value, TRUE/FALSE from checkbox input

is_on_disk
Logical, whether df represented by df_label was on disk or from interactive R use

out_path
reactive, List, with character strings providing directory paths and file names for saving/reading in code output