Normalise all columns of a salary table

Wrapper (orchestrator) function that sequentially applies all normalisation rules specific to the salary dataset (salary_tbl).

normalize_all(data)

Arguments

data

The data.frame (salary_tbl) to normalise.

Value

The fully normalised data.frame, ready for analysis.

Details

This function transforms the following columns:

Location (company_location, employee_residence): ISO2 normalisation and creation of regional groupings.

Job title (job_title): Standardisation of job titles.

Remote work (remote_ratio): Numeric cleaning of the remote work ratio.

Company size (company_size): Conversion to an ordered factor.

Employment type (employment_type): Standardisation.

Experience (experience_level): Conversion to an ordered factor.

Note

This function depends on global mapping variables (e.g. mapping_total, levels_iso2, size_mapping, etc.) that must be loaded in the environment before execution.

See also

Examples

if (FALSE) { # \dontrun{
# This example does not run automatically because it requires
# the mapping objects (mapping_total, levels_iso2, etc.)
# to be present in the global environment.

df_clean <- normalize_all(salary_tbl_raw)
} # }