Normalize a vector to a factor. (use with "df$col <- normalize_to_factor(df$col, mapping, levels)")

Transforme un vecteur brut en facteur en appliquant une table de correspondance (mapping). Les valeurs non trouvées ou manquantes sont remplacées par "Unknown".

normalize_factor(vector, mapping, levels_in_factor, ordered_factor = FALSE)

Arguments

vector

Le vecteur brut à normaliser.

mapping

Un vecteur nommé (clé-valeur) indiquant la transformation (ex: c("old_val" = "new_val")).

levels_in_factor

Un vecteur contenant tous les niveaux valides pour le facteur final.

ordered_factor

Logique (défaut FALSE). Si TRUE, crée un facteur ordonné.

Value

Un vecteur de type factor normalisé.

Details

Cette fonction est utile pour standardiser des colonnes catégorielles mal formatées. Elle effectue les opérations suivantes :

Remplace les valeurs du vecteur selon le mapping fourni.

Remplace les NA par la chaîne "Unknown".

Convertit le résultat en factor avec les niveaux (levels) imposés.

See also

Examples

raw_data <- c("FR", "US", "UK", "FR", "XX")

# Table de correspondance
map_countries <- c(
  "FR" = "France",
  "US" = "United States",
  "UK" = "United Kingdom"
)

# Niveaux autorisés
valid_levels <- c("France", "United States", "United Kingdom", "Unknown")

# Application
normalize_factor(raw_data, map_countries, valid_levels)
#>             FR             US             UK             FR           <NA> 
#>         France  United States United Kingdom         France        Unknown 
#> Levels: France United States United Kingdom Unknown