Exporter les données en CSV

Sauvegarde un data.frame dans un fichier CSV sur le disque. Cette fonction gère automatiquement la création du dossier cible, l'ajout de l'extension .csv et le choix du séparateur.

export_csv(
  data,
  path = "exports",
  filename = "cleaned_data",
  sep = ";",
  row.names = FALSE,
  overwrite = TRUE,
  verbose = TRUE
)

Arguments

data

Le data.frame à exporter.

path

Chemin du dossier de destination (défaut "exports"). Le dossier sera créé s'il n'existe pas.

filename

Nom du fichier. L'extension .csv est ajoutée automatiquement si nécessaire (défaut "cleaned_data").

sep

Caractère séparateur de colonnes (défaut ";").

row.names

Logique. Faut-il inclure les noms de lignes dans le fichier ? (défaut FALSE).

overwrite

Logique. Si FALSE, lève une erreur si le fichier existe déjà. Si TRUE (défaut), l'écrase.

verbose

Logique. Affiche un message avec le chemin complet du fichier après écriture (défaut TRUE).

Value

Renvoie de manière invisible (invisible()) le chemin complet du fichier créé.

Details

La fonction simplifie l'exportation en effectuant les étapes suivantes :

Vérifie que l'objet data est bien un data.frame.

Crée le dossier path récursivement s'il n'existe pas encore.

Ajoute l'extension .csv au nom du fichier si elle est absente.

Choisit la fonction d'écriture optimisée selon le séparateur :

Pour sep = ";" : utilise utils::write.csv2 (format européen).

Pour sep = "," : utilise utils::write.csv (format US).

Autres : utilise utils::write.table.

Examples

# Données de test
df_test <- data.frame(
  id = 1:3,
  nom = c("Alice", "Bob", "Charlie"),
  score = c(10.5, 15.2, 8.0)
)

# Création d'un dossier temporaire pour ne pas polluer votre disque
tmp_folder <- tempfile() # On utilise un nom aléatoire

# 1. Export simple (point-virgule par défaut)
path_file <- export_csv(df_test, path = tmp_folder, filename = "test_export")
#> Fichier écrit: C:\Users\margo\AppData\Local\Temp\RtmpCesNfG\file6f34373e1df5\test_export.csv

# Vérification que le fichier existe
file.exists(path_file)
#> [1] TRUE

# 2. Export avec virgule et sans écraser (devrait générer une erreur si on relance)
try({
  export_csv(df_test, path = tmp_folder, filename = "test_export", sep = ",", overwrite = FALSE)
})
#> Error in export_csv(df_test, path = tmp_folder, filename = "test_export",  : 
#>   Le fichier existe déjà : C:\Users\margo\AppData\Local\Temp\RtmpCesNfG\file6f34373e1df5/test_export.csv

# Nettoyage du dossier temporaire
unlink(tmp_folder, recursive = TRUE)