Wrapper for adjusting axis lims and hiding traces

Wrapper for adjusting axis lims and hiding traces

handle_restyle_traces(
  source_id,
  session,
  dframe,
  scaling = 0.05,
  xvar,
  yvar,
  trace_map,
  max_id_group_trace,
  input_sel_rows,
  flush = TRUE
)

Arguments

source_id

character, plotly source id

session

session object

dframe

data frame/tibble (grouped/ungrouped)

scaling

numeric, 1 +/- scaling applied to x lims for xvar and yvar

xvar

character, name of xvar, must be in dframe

yvar

character, name of yvar, must be in dframe

trace_map

matrix, with columns for trace name (col 1) and trace id (col 2)

max_id_group_trace

numeric, max id of plotly trace from original data (not outlier traces)

input_sel_rows

numeric, input from DT grouptable

flush

character, plotlyProxy settings

Value

Used for it's side effect - no return